edit and remove functionality
Can now edit and delete transactions and categories as well as delete users. You need to delete your databases and reload them for it to work. I will create documentation and update my current documents later when I get a chance. get_transactions and get_categories now return a list of lists. The first item in the sublist is the the transaction or category id.